+++ enrollment-service/src/main/java/com/project/enrollmentservice/model/Enrollment.java
@@ -18,7 +18,7 @@ public class Enrollment {
     private Long enrollmentID;
 
     @NotNull(message = "Student ID is required")
-    private int studentID;
+    private Long studentID;
 
     @NotNull(message = "Offering ID is required")
     private Long offeringID;
